backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
httpx==0.24.1
python-dotenv==1.0.0
websockets==11.0.3